refactor: split forward_block.rs and qwen35_dense.rs under 500-LoC cap
Extract the fc-projection step of BlockDiffusionDraftHead::forward_block
into forward_block_step0.rs, and the LinearAttention (GDN SSM) layer
loading arm of Qwen35DenseWeightLoader::load_layers into a new
qwen35_dense/load_linear_attn.rs submodule, following the existing
args-bundle and variant-dispatch module patterns used elsewhere in the
crate. No behavior change.
